move mocking of jira-client to mockClients.ts
